+++ ChangeLog.md
@@ -6,7 +6,8 @@ Incoming
 
 * Breaking changes
     * Removed `Pattern::bind()`. Use `Pattern::inject()` or `Pattern::template()->literal()`.
-    * Removed `Pattern::template()->bind()`..
+    * Removed `Pattern::pcre()->bind()`.
+    * Removed `Pattern::template()->bind()`.
     * Removed `Pattern::template()->builder()`. Use `Pattern::template()`.
 * Bug fixes
     * Correct type-error in `ValidPattern.isValid()` on PHP 8.1.